+++ R/writeODS.R
@@ -1,4 +1,4 @@
-.zip_tmp_to_path <- function(temp_ods_dir, path, overwrite) {
+.zip_tmp_to_path <- function(temp_ods_dir, path, overwrite = TRUE) {
     wd <- getwd()
     on.exit(setwd(wd), add = TRUE)
     setwd(temp_ods_dir)
@@ -122,9 +122,7 @@
 #' @param update logical, TRUE indicates that the sheet with sheet_name in the existing file (path) should be updated with the content of x. If a sheet with sheet_name does not exist, an exception is thrown. Please also note that writing is slower if TRUE. Default is FALSE.
 #' @param row_names logical, TRUE indicates that row names of x are to be included in the sheet. Default is FALSE.
 #' @param col_names logical, TRUE indicates that column names of x are to be included in the sheet. Default is FALSE.
-#' @param na_as_string logical, TRUE indicates that NAs are written as string. Default is `option("write_ods_na")` (will change to TRUE in the next version).
-#' @param verbose logical, deprecated
-#' @param overwrite logical, deprecated.
+#' @param na_as_string logical, TRUE indicates that NAs are written as string.
 #' @return An ODS file written to the file path location specified by the user. The value of \code{path} is also returned invisibly.
 #' @author Detlef Steuer <steuer@@hsu-hh.de>, Thomas J. Leeper <thosjleeper@@gmail.com>, John Foster <john.x.foster@@nab.com.au>, Chung-hong Chan <chainsawtiney@@gmail.com>
 #' @examples
@@ -135,26 +133,16 @@
 #' write_ods(PlantGrowth, "mtcars.ods", append = TRUE, sheet = "plant")
 #' }
 #' @export
-write_ods <- function(x, path, sheet = "Sheet1", append = FALSE, update = FALSE, row_names = FALSE, col_names = TRUE, na_as_string = getOption("write_ods_na", default = FALSE), verbose = NULL, overwrite = NULL) {
+write_ods <- function(x, path, sheet = "Sheet1", append = FALSE, update = FALSE, row_names = FALSE, col_names = TRUE, na_as_string = TRUE) {
     ## setup temp directory
     ## one can't just use tempdir() because it is the same in the same session
     temp_ods_dir <- file.path(tempdir(), stringi::stri_rand_strings(1, 20, pattern = "[A-Za-z0-9]"))
     dir.create(temp_ods_dir)
     on.exit(unlink(temp_ods_dir))
-    if (!is.null(overwrite)) {
-        warning("overwrite is deprecated. Future versions will always set it to TRUE.")
-    } else {
-        overwrite <- TRUE
-    }
-    if (!is.null(verbose)) {
-        warning("verbose is deprecated. Future versions will always set it to FALSE.")
-    } else {
-        verbose <- FALSE
-    }
     if (!is.data.frame(x)) {
         stop("x must be a data.frame.", call. = FALSE)
     }
-    if (!file.exists(path) | (!append & !update)) {
+    if (!file.exists(path) || (!append && !update)) {
         .vfwrite_ods(x = x, temp_ods_dir = temp_ods_dir, sheet = sheet, row_names = row_names, col_names = col_names, na_as_string = na_as_string)
     } else {
         ## The file must be there.
@@ -163,18 +151,18 @@ write_ods <- function(x, path, sheet = "Sheet1", append = FALSE, update = FALSE,
         content <- xml2::read_xml(contentfile)
         spreadsheet_node <- xml2::xml_children(xml2::xml_children(content)[[which(!is.na(xml2::xml_find_first(xml2::xml_children(content),"office:spreadsheet")))]])[[1]]
         sheet_node <- .find_sheet_node_by_sheet(spreadsheet_node, sheet)
-        if ((!is.null(sheet_node) & append & !update) | (!is.null(sheet_node) & !update)) {
+        if ((!is.null(sheet_node) && append && !update) || (!is.null(sheet_node) && !update)) {
             ## Sheet exists so we cannot append
             stop(paste0("Sheet ", sheet, " exists. Set update to TRUE is you want to update this sheet."), call. = FALSE)
         }
-        if (is.null(sheet_node) & update) {
+        if (is.null(sheet_node) && update) {
             stop(paste0("Sheet ", sheet, " does not exist. Cannot update."), call. = FALSE)
         }
-        if (!is.null(sheet_node) & update) {
+        if (!is.null(sheet_node) && update) {
             ## clean up the sheet
             xml2::xml_remove(xml2::xml_children(sheet_node)[2:length(xml2::xml_children(sheet_node))])
         }
-        if (is.null(sheet_node) & append) {
+        if (is.null(sheet_node) && append) {
             ## Add a new sheet
             sheet_node <- xml2::xml_add_child(spreadsheet_node, .silent_add_sheet_node(sheet))
         }
@@ -185,6 +173,6 @@ write_ods <- function(x, path, sheet = "Sheet1", append = FALSE, update = FALSE,
         xml2::write_xml(content, contentfile)
     }
     ## zip up ODS archive
-    .zip_tmp_to_path(temp_ods_dir, path, overwrite)
+    .zip_tmp_to_path(temp_ods_dir, path)
     invisible(path)
 }